Source, SourceType, and Index are all optional
Splunk will pick reasonable defaults if the first two are omitted.

A default index is always assigned when a token in created in Splunk.
However, the JSON request must not refer to a blank index if this
default is to be used.  Without the json omitempty, the log entry is
written to the index "main" where you might not think to look.