Fix memory corruption in io.Writer interface
When hooked up to logrus and running multiple log.Info() calls (for example),
the bytearray being passed to Write() is reused in the successive calls.  As
this reference was being passed down through the channel, it resulted in
corrupted data and JSON serialization errors.

We fix that here by always making a copy before sending to the channel.